unit and integration tests run separated, install and deploy plugin configured
