+++ pom.xml
@@ -20,6 +20,8 @@
         <devkit.studio.package.skip>false</devkit.studio.package.skip>
         <connector.test.version>2.0.6</connector.test.version>
         <harmcrest.version>1.3</harmcrest.version>
+        <skip.unit.tests>false</skip.unit.tests>
+        <skip.integration.tests>false</skip.integration.tests>
     </properties>
 
     <dependencies>
@@ -134,13 +136,42 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.17</version>
+                <version>2.19.1</version>
                 <configuration>
+                    <enableAssertions>false</enableAssertions>
+                    <failIfNoTests>false</failIfNoTests>
+                    <skipTests>${skip.unit.tests}</skipTests>
                     <excludes>
-                        <exclude>**/CassandraDBObjectStoreFunctionalTestCase.java</exclude>
+                        <exclude>**/functional/**.*.java</exclude>
+                        <exclude>**/system/**.*.java</exclude>
                     </excludes>
                 </configuration>
             </plugin>
+            <plugin>
+                <artifactId>maven-failsafe-plugin</artifactId>
+                <version>2.20.1</version>
+                <configuration>
+                    <enableAssertions>false</enableAssertions>
+                    <failIfNoTests>false</failIfNoTests>
+                    <skipTests>${skip.integration.tests}</skipTests>
+                    <includes>
+                        <include>**/functional/**.*.java</include>
+                        <include>**/system/**.*.java</include>
+                        <include>**/runner/*TestSuite.java</include>
+                        <include>**/*IT.java</include>
+                    </includes>
+                </configuration>
+                <executions>
+                    <execution>
+                        <id>integration-test</id>
+                        <goals>
+                            <goal>integration-test</goal>
+                            <goal>verify</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+
             <plugin>
 			    <artifactId>maven-source-plugin</artifactId>
 			    <version>2.1.2</version>
@@ -165,6 +196,20 @@
                 </executions>
             </plugin>
         </plugins>
+        <pluginManagement>
+            <plugins>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-install-plugin</artifactId>
+                    <version>2.5.2</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-deploy-plugin</artifactId>
+                    <version>2.8.2</version>
+                </plugin>
+            </plugins>
+        </pluginManagement>
     </build>
 
     <scm>
