+++ src/test/java/org/mule/modules/cassandradb/automation/unit/ConnectionUtilTestCase.java
@@ -12,7 +12,7 @@ import java.util.Map;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-public class ConnectionUtilTestCases {
+public class ConnectionUtilTestCase {
     public static final String CASSANDRA_NODE_DEFAULT_PORT = "9042";
 
     private Map<String, String> nodes = new HashMap<>();
